Tested other auth functions.  Rejigged error response. Fairly major reshuffle of internals.
