+++ Enums/DirectXVersion.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ISBoxerEVELauncher.Enums
+{
+    public enum DirectXVersion
+    {
+        Default,
+        dx9,
+        dx11,
+    }
+}
