+++ Enums/RelatedExecutable.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ISBoxerEVELauncher.Enums
+{
+    public enum RelatedExecutable
+    {
+        InvalidGameProfile,
+        Other,
+        EXEFile,
+        EVELauncher,
+        ISBoxerEVELauncher,
+        InnerSpace,
+    }
+}
