+++ Extensions/HttpWebRequestExtension.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Net;
 using System.Text;
 using System.Threading.Tasks;
+using ISBoxerEVELauncher.Security;
 
 namespace ISBoxerEVELauncher.Extensions
 {
