+++ ISBoxerEVELauncher.csproj
@@ -45,6 +45,9 @@
     <Reference Include="Costura, Version=3.3.2.0, Culture=neutral, PublicKeyToken=9919ef960d84173d, processorArchitecture=MSIL">
       <HintPath>packages\Costura.Fody.3.3.2\lib\net40\Costura.dll</HintPath>
     </Reference>
+    <Reference Include="HtmlAgilityPack, Version=1.8.14.0, Culture=neutral, PublicKeyToken=bd319b19eaf3b43a, processorArchitecture=MSIL">
+      <HintPath>packages\HtmlAgilityPack.1.8.14\lib\Net45\HtmlAgilityPack.dll</HintPath>
+    </Reference>
     <Reference Include="Microsoft.IdentityModel.Logging, Version=5.4.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>packages\Microsoft.IdentityModel.Logging.5.4.0\lib\net45\Microsoft.IdentityModel.Logging.dll</HintPath>
     </Reference>
@@ -77,20 +80,23 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="CookieStorage.cs" />
-    <Compile Include="EVECharacter.cs" />
+    <Compile Include="Enums\DirectXVersion.cs" />
+    <Compile Include="Enums\RelatedExecutable.cs" />
+    <Compile Include="Web\CookieStorage.cs" />
+    <Compile Include="Games\EVE\EVECharacter.cs" />
     <Compile Include="Extensions\HttpWebResponseExtension.cs" />
     <Compile Include="Extensions\ProcessExtension.cs" />
     <Compile Include="Extensions\UriExtension.cs" />
     <Compile Include="Extensions\HttpWebRequestExtension.cs" />
     <Compile Include="Extensions\StringExtension.cs" />
     <Compile Include="Extensions\WebResponseExtension.cs" />
-    <Compile Include="KeyTransmitter.cs" />
+    <Compile Include="Security\KeyTransmitter.cs" />
     <Compile Include="Launchers.cs" />
-    <Compile Include="LoginResultEnum.cs" />
-    <Compile Include="SecureBytesWrapper.cs" />
-    <Compile Include="Utility\Crypto.cs" />
-    <Compile Include="Utility\WebReqResp.cs" />
+    <Compile Include="Enums\LoginResult.cs" />
+    <Compile Include="Security\SecureBytesWrapper.cs" />
+    <Compile Include="Security\SHA256.cs" />
+    <Compile Include="Web\RequestResponse.cs" />
+    <Compile Include="Web\Response.cs" />
     <Compile Include="Windows\AddCharacterWindow.xaml.cs">
       <DependentUpon>AddCharacterWindow.xaml</DependentUpon>
     </Compile>
@@ -112,14 +118,14 @@
     <Compile Include="Windows\EVEEULAWindow.xaml.cs">
       <DependentUpon>EVEEULAWindow.xaml</DependentUpon>
     </Compile>
-    <Compile Include="InnerSpaceGameProfile.cs" />
+    <Compile Include="InnerSpace\InnerSpaceGameProfile.cs" />
     <Compile Include="Windows\LaunchProgressWindow.xaml.cs">
       <DependentUpon>LaunchProgressWindow.xaml</DependentUpon>
     </Compile>
     <Compile Include="Windows\MasterKeyEntryWindow.xaml.cs">
       <DependentUpon>MasterKeyEntryWindow.xaml</DependentUpon>
     </Compile>
-    <Compile Include="SecureStringWrapper.cs" />
+    <Compile Include="Security\SecureStringWrapper.cs" />
     <Compile Include="Windows\SecurityWarningWindow.xaml.cs">
       <DependentUpon>SecurityWarningWindow.xaml</DependentUpon>
     </Compile>
@@ -177,11 +183,11 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="EVEAccount.cs" />
+    <Compile Include="Games\EVE\EVEAccount.cs" />
     <Compile Include="Windows\EVELogin.xaml.cs">
       <DependentUpon>EVELogin.xaml</DependentUpon>
     </Compile>
-    <Compile Include="InnerSpaceSettings.cs" />
+    <Compile Include="InnerSpace\InnerSpaceSettings.cs" />
     <Compile Include="Windows\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
@@ -242,6 +248,7 @@
   <ItemGroup>
     <Resource Include="FodyWeavers.xml" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="packages\Fody.4.0.0\build\Fody.targets" Condition="Exists('packages\Fody.4.0.0\build\Fody.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
