+++ Windows/LaunchProgressWindow.xaml.cs
@@ -14,6 +14,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 using ISBoxerEVELauncher.Launchers;
+using ISBoxerEVELauncher.Enums;
 
 namespace ISBoxerEVELauncher.Windows
 {
