+++ packages.config
@@ -2,6 +2,7 @@
 <packages>
   <package id="Costura.Fody" version="3.3.2" targetFramework="net45" />
   <package id="Fody" version="4.0.0" targetFramework="net45" developmentDependency="true" />
+  <package id="HtmlAgilityPack" version="1.8.14" targetFramework="net45" />
   <package id="Microsoft.IdentityModel.Logging" version="5.4.0" targetFramework="net45" />
   <package id="Microsoft.IdentityModel.Tokens" version="5.4.0" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net45" />
